wip; restructuring cluster tests to reduce CI duration
